Move date formatting helpers to lib
